Added error codes and edge cases

+++ index.js
@@ -15,7 +15,28 @@ server.get("/api/posts", (req, res) => {
     .then(posts => {
       res.json(posts);
     })
-    .catch(err => res.send(err));
+    .catch(err =>
+      res.status(500).json({
+        error: "The posts information could not be retrieved."
+      })
+    );
+});
+
+server.get("/api/posts/:id", (req, res) => {
+  db.findById(req.params.id)
+    .then(post => {
+      if (post.length > 0) {
+        res.json(post);
+      } else
+        res.status(404).json({
+          message: "The post with the specified ID does not exist."
+        });
+    })
+    .catch(err =>
+      res
+        .status(500)
+        .json({ error: "The post information could not be retrieved." })
+    );
 });
 
 server.post("/api/posts", (req, res) => {
@@ -24,37 +45,63 @@ server.post("/api/posts", (req, res) => {
   if (!title || !contents) {
     return res
       .status(400)
-      .send({ Error: `Please provide title and contents for the post.` });
+      .json({ error: "Please provide title and contents for the post." });
   }
   db.insert(newPost)
     .then(postId => {
       const { id } = postId;
       db.findById(id).then(post => {
-        console.log(post);
         res.status(201).json(post);
       });
     })
-    .catch(err => console.error(err));
+    .catch(err =>
+      res.status(500).json({
+        error: "There was an error while saving the post to the database."
+      })
+    );
 });
 
 server.delete("/api/posts/:id", (req, res) => {
   const { id } = req.params;
+  if (post.length < 1) {
+    return res.status(404).json({
+      message: "The post with the specified ID does not exist."
+    });
+  }
   db.remove(id)
     .then(removedPost => {
       res.status(200).json(removedPost);
     })
-    .catch(err => console.error(err));
+    .catch(err =>
+      res.status(500).json({
+        error: "The post could not be removed."
+      })
+    );
 });
 
 server.put("/api/posts/:id", (req, res) => {
   const { id } = req.params;
   const { title, contents } = req.body;
   const newPost = { title, contents };
+  if (post.length < 1) {
+    return res.status(404).json({
+      message: "The post with the specified ID does not exist."
+    });
+  }
+  if (!title || !contents) {
+    return res
+      .status(400)
+      .json({ error: "Please provide title and contents for the post." });
+  }
   db.update(id, newPost)
     .then(post => {
       res.status(200).json(post);
     })
-    .catch(err => console.error(err));
+    .catch(err =>
+      res.status(500).json({
+        error: "The post information could not be modified.."
+      })
+    );
 });
 
 const port = 8000;
